+++ src/util/context/TaskContext.tsx
@@ -1,6 +1,7 @@
 import { collection, getDoc, doc } from 'firebase/firestore';
 import React, { createContext, useContext, useState } from 'react';
 import { db } from '../../../firebaseConfig';
+
 interface Tasks {
   taskTitle?: string;
   taskDescription?: string;
